\name{ClusterRankTest-package}
\alias{ClusterRankTest-package}
\alias{ClusterRankTest}
\docType{package}
\title{Rank Tests for Clustered Data}
\description{This package consists of methods that compute rank based tests for clustered data.}
\details{
\tabular{ll}{
Package: \tab ClusterRankTest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-04-21\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Sandipan Dutta, Somnath Datta

Maintainer: Sandipan Dutta <sandipan.dutta07@gmail.com>

}
