\name{dataset}
\alias{dataset}
\docType{data}
\title{dataset}
\description{
The dataset has three columns. The first two columns will be 'Cluster ID' and 'X'(outcome). The third column will be either 'Group Indicator' (for rank-sum tests) or 'Y'(paired outcome for signed-rank tests).
}
\usage{data("dataset")}
